migrate show model
